tests/conformance: refresh protocol suites for updated APIs + stricter validation
Bulk update of the conformance test surface to match the runtime,
network, and raptorq changes landing in this release. Tests stay on the
same RFC scope; all churn is shape/import alignment and coverage added
for the new stricter error paths.

HTTP/2 (h2_connect, h2_alpn_negotiation_rfc7540, h2_rst_stream_ping_rfc9113,
hpack_rfc7541, hpack_table_size, h2_goaway_graceful_shutdown,
hpack_eviction, hpack_huffman):
- Drop vestigial asupersync imports (Bytes, Cx, Connection, Frame,
  FrameHeader, FrameType, Settings, TestTimeGetter, SystemTime) that no
  longer participate in the harness after the h2/hpack rework.
- Switch to a local ConnectTestTimeGetter so deterministic time is owned
  by the harness rather than threaded through the runtime layer. Mirror
  helper shapes used by the other h2 conformance suites.
- HPACK: adopt split_to / set_allowed_table_size in place of the removed
  Bytes::advance / Decoder::set_max_table_size call sites; cover stale
  eviction and Huffman paths so the audit findings stay regression-bound.

HTTP/3 (h3_settings, trace_replay_idempotency_metamorphic):
- Align with the new H3EndpointRole + pseudo-header validation surface
  so invalid :method and CR/LF :authority values are asserted to fail
  construction and decode.

HTTP/1.1 (h1_rfc9112/chunked_encoding_tests, golden_test_vectors, harness,
conformance_h1_expect):
- Minor harness plumbing to pick up the updated cancel-checkpoint path
  through tcp/stream.rs.

gRPC (grpc_trailer_forwarding_rfc9113):
- Update expected trailer forwarding behavior to the new interceptor +
  health watch cancellation semantics.

Kafka (kafka_offsets, kafka_record_batch_v2/{golden_tests,test_vectors}):
- Sync with the messaging/kafka.rs offset + record-batch shape changes.

MySQL / Postgres (mysql_stmt_prepare_execute, postgres_logical_replication):
- Align prepared-statement and logical-replication tests with the
  postgres strict bool decoding + prefer-mode cancel fidelity changes.

QUIC / TLS / WebSocket (quic_retry_rfc9000, tls_0rtt_replay_rfc8446,
websocket_extension_negotiation_rfc6455, conformance_websocket):
- Thread the new cancel-checkpoint surface through these suites; no
  protocol behavior change.

Aggregator / broadcast / cancel_dag_determinism / tcp_accept /
race_loser_drain_metamorphic:
- Pick up the updated scheduler + race-loser-drain semantics.

Obligation invariants (obligation_invariants/*, obligation_lifecycle_
metamorphic):
- Consume the NoLeakProver closed-region counterexample and the ledger
  id generation split introduced in src/obligation/ledger.rs.

Shared scaffolding (conformance/mod.rs, conformance_io_uring_buffer_pool,
conformance_diagnostics_endpoints):
- Drop dead imports and align the common harness exports with the
  updated observability + diagnostics surfaces.

+++ tests/conformance/kafka_record_batch_v2/golden_tests.rs
@@ -218,7 +218,7 @@ mod tests {
                 Some(format!("offset-value-{}", i).into_bytes()),
             )
             .with_timestamp_delta(i * 5)
-            .with_offset_delta(i);
+            .with_offset_delta(i as i32);
 
             batch.add_record(record);
         }
